+++ src/hex/impls.rs
@@ -1,7 +1,10 @@
 use super::Hex;
 use std::{
     iter::{Product, Sum},
-    ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Rem, RemAssign, Sub, SubAssign},
+    ops::{
+        Add, AddAssign, BitAnd, BitOr, BitXor, Div, DivAssign, Mul, MulAssign, Neg, Rem, RemAssign,
+        Shl, Shr, Sub, SubAssign,
+    },
 };
 
 impl Add<Self> for Hex {
@@ -19,8 +22,8 @@ impl Add<i32> for Hex {
     #[inline]
     fn add(self, rhs: i32) -> Self::Output {
         Self {
-            x: self.x + rhs,
-            y: self.y + rhs,
+            x: self.x.add(rhs),
+            y: self.y.add(rhs),
         }
     }
 }
@@ -28,26 +31,26 @@ impl Add<i32> for Hex {
 impl AddAssign for Hex {
     #[inline]
     fn add_assign(&mut self, rhs: Self) {
-        *self = *self + rhs;
+        *self = self.add(rhs);
     }
 }
 
 impl AddAssign<i32> for Hex {
     #[inline]
     fn add_assign(&mut self, rhs: i32) {
-        *self = *self + rhs;
+        *self = self.add(rhs);
     }
 }
 
 impl Sum for Hex {
     fn sum<I: Iterator<Item = Self>>(iter: I) -> Self {
-        iter.fold(Self::ZERO, |a, b| a + b)
+        iter.fold(Self::ZERO, Self::const_add)
     }
 }
 
 impl<'a> Sum<&'a Self> for Hex {
     fn sum<I: Iterator<Item = &'a Self>>(iter: I) -> Self {
-        iter.copied().sum()
+        iter.fold(Self::ZERO, |a, &b| Self::const_add(a, b))
     }
 }
 
@@ -66,8 +69,8 @@ impl Sub<i32> for Hex {
     #[inline]
     fn sub(self, rhs: i32) -> Self::Output {
         Self {
-            x: self.x - rhs,
-            y: self.y - rhs,
+            x: self.x.sub(rhs),
+            y: self.y.sub(rhs),
         }
     }
 }
@@ -75,14 +78,14 @@ impl Sub<i32> for Hex {
 impl SubAssign for Hex {
     #[inline]
     fn sub_assign(&mut self, rhs: Self) {
-        *self = *self - rhs;
+        *self = self.sub(rhs);
     }
 }
 
 impl SubAssign<i32> for Hex {
     #[inline]
     fn sub_assign(&mut self, rhs: i32) {
-        *self = *self - rhs;
+        *self = self.sub(rhs);
     }
 }
 
@@ -92,8 +95,8 @@ impl Mul<Self> for Hex {
     #[inline]
     fn mul(self, rhs: Self) -> Self::Output {
         Self {
-            x: self.x * rhs.x,
-            y: self.y * rhs.y,
+            x: self.x.mul(rhs.x),
+            y: self.y.mul(rhs.y),
         }
     }
 }
@@ -104,8 +107,8 @@ impl Mul<i32> for Hex {
     #[inline]
     fn mul(self, rhs: i32) -> Self::Output {
         Self {
-            x: self.x * rhs,
-            y: self.y * rhs,
+            x: self.x.mul(rhs),
+            y: self.y.mul(rhs),
         }
     }
 }
@@ -123,33 +126,33 @@ impl Mul<f32> for Hex {
 impl MulAssign for Hex {
     #[inline]
     fn mul_assign(&mut self, rhs: Self) {
-        *self = *self * rhs;
+        *self = self.mul(rhs);
     }
 }
 
 impl MulAssign<i32> for Hex {
     #[inline]
     fn mul_assign(&mut self, rhs: i32) {
-        *self = *self * rhs;
+        *self = self.mul(rhs);
     }
 }
 
 impl MulAssign<f32> for Hex {
     #[inline]
     fn mul_assign(&mut self, rhs: f32) {
-        *self = *self * rhs;
+        *self = self.mul(rhs);
     }
 }
 
 impl Product for Hex {
     fn product<I: Iterator<Item = Self>>(iter: I) -> Self {
-        iter.fold(Self::ZERO, |a, b| a * b)
+        iter.fold(Self::ONE, Self::mul)
     }
 }
 
 impl<'a> Product<&'a Self> for Hex {
     fn product<I: Iterator<Item = &'a Self>>(iter: I) -> Self {
-        iter.copied().product()
+        iter.fold(Self::ONE, |a, &b| Self::mul(a, b))
     }
 }
 
@@ -159,8 +162,8 @@ impl Div<Self> for Hex {
     #[inline]
     fn div(self, rhs: Self) -> Self::Output {
         Self {
-            x: self.x / rhs.x,
-            y: self.y / rhs.y,
+            x: self.x.div(rhs.x),
+            y: self.y.div(rhs.y),
         }
     }
 }
@@ -195,21 +198,21 @@ impl Div<f32> for Hex {
 impl DivAssign for Hex {
     #[inline]
     fn div_assign(&mut self, rhs: Self) {
-        *self = *self / rhs;
+        *self = self.div(rhs);
     }
 }
 
 impl DivAssign<i32> for Hex {
     #[inline]
     fn div_assign(&mut self, rhs: i32) {
-        *self = *self / rhs;
+        *self = self.div(rhs);
     }
 }
 
 impl DivAssign<f32> for Hex {
     #[inline]
     fn div_assign(&mut self, rhs: f32) {
-        *self = *self / rhs;
+        *self = self.div(rhs);
     }
 }
 
@@ -234,14 +237,14 @@ impl Rem<i32> for Hex {
 impl RemAssign for Hex {
     #[inline]
     fn rem_assign(&mut self, rhs: Self) {
-        *self = *self % rhs;
+        *self = self.rem(rhs);
     }
 }
 
 impl RemAssign<i32> for Hex {
     #[inline]
     fn rem_assign(&mut self, rhs: i32) {
-        *self = *self % rhs;
+        *self = self.rem(rhs);
     }
 }
 
@@ -251,5 +254,214 @@ impl Neg for Hex {
     #[inline]
     fn neg(self) -> Self::Output {
         self.const_neg()
+    }
+}
+
+impl BitAnd for Hex {
+    type Output = Self;
+    #[inline]
+    fn bitand(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x.bitand(rhs.x),
+            y: self.y.bitand(rhs.y),
+        }
+    }
+}
+
+impl BitOr for Hex {
+    type Output = Self;
+    #[inline]
+    fn bitor(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x.bitor(rhs.x),
+            y: self.y.bitor(rhs.y),
+        }
+    }
+}
+
+impl BitXor for Hex {
+    type Output = Self;
+    #[inline]
+    fn bitxor(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x.bitxor(rhs.x),
+            y: self.y.bitxor(rhs.y),
+        }
+    }
+}
+
+impl BitAnd<i32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn bitand(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x.bitand(rhs),
+            y: self.y.bitand(rhs),
+        }
+    }
+}
+
+impl BitOr<i32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn bitor(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x.bitor(rhs),
+            y: self.y.bitor(rhs),
+        }
+    }
+}
+
+impl BitXor<i32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn bitxor(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x.bitxor(rhs),
+            y: self.y.bitxor(rhs),
+        }
+    }
+}
+
+impl Shl<i8> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: i8) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs),
+            y: self.y.shl(rhs),
+        }
+    }
+}
+
+impl Shr<i8> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shr(self, rhs: i8) -> Self::Output {
+        Self {
+            x: self.x.shr(rhs),
+            y: self.y.shr(rhs),
+        }
+    }
+}
+
+impl Shl<i16> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: i16) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs),
+            y: self.y.shl(rhs),
+        }
+    }
+}
+
+impl Shr<i16> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shr(self, rhs: i16) -> Self::Output {
+        Self {
+            x: self.x.shr(rhs),
+            y: self.y.shr(rhs),
+        }
+    }
+}
+
+impl Shl<i32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs),
+            y: self.y.shl(rhs),
+        }
+    }
+}
+
+impl Shr<i32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shr(self, rhs: i32) -> Self::Output {
+        Self {
+            x: self.x.shr(rhs),
+            y: self.y.shr(rhs),
+        }
+    }
+}
+
+impl Shl<u8> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: u8) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs),
+            y: self.y.shl(rhs),
+        }
+    }
+}
+
+impl Shr<u8> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shr(self, rhs: u8) -> Self::Output {
+        Self {
+            x: self.x.shr(rhs),
+            y: self.y.shr(rhs),
+        }
+    }
+}
+
+impl Shl<u16> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: u16) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs),
+            y: self.y.shl(rhs),
+        }
+    }
+}
+
+impl Shr<u16> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shr(self, rhs: u16) -> Self::Output {
+        Self {
+            x: self.x.shr(rhs),
+            y: self.y.shr(rhs),
+        }
+    }
+}
+
+impl Shl<u32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: u32) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs),
+            y: self.y.shl(rhs),
+        }
+    }
+}
+
+impl Shr<u32> for Hex {
+    type Output = Self;
+    #[inline]
+    fn shr(self, rhs: u32) -> Self::Output {
+        Self {
+            x: self.x.shr(rhs),
+            y: self.y.shr(rhs),
+        }
+    }
+}
+
+impl Shl for Hex {
+    type Output = Self;
+    #[inline]
+    fn shl(self, rhs: Self) -> Self::Output {
+        Self {
+            x: self.x.shl(rhs.x),
+            y: self.y.shl(rhs.y),
+        }
     }
 }